Apply feedback from PR reviews: fix null pointer, improve password validation, create reusable components
